Parameterize CI and compose file (#16)
* Parameterize CI and compose file

* Install Docker on the arm hosts

* Update method for installing Docker on arm runners

* Fix ssh timing issue

* Fix permissions on .ssh directory

* Use healthchecks to work around ssh key gen race conditions

* Try standard runners and remove comments

* Remove docker install from standard arm runner jobs

* Fix bug in master entrypoint slurm cpus setting

* Remove nocopy mount option

* Wait for ssh to be ready on nodes

* Fix indentation typo in node health check

* Force compose to wait for slurmnode health (ssh readiness) before completion

+++ frontend/docker-entrypoint.sh
@@ -1,13 +1,21 @@
 #!/bin/bash
 
-export SLURM_CPUS_ON_NODE=$(cat /proc/cpuinfo | grep processor | wc -l)
+export SLURM_CPUS_ON_NODE=${SLURM_CPUS_ON_NODE:-$(cat /proc/cpuinfo | grep processor | wc -l)}
 sudo sed -i "s/REPLACE_IT/${SLURM_CPUS_ON_NODE}/g" /etc/slurm/slurm.conf
 
 sudo service munge start
 sudo ssh-keygen -A
 sudo service ssh start
 
-ssh-keygen -t rsa -f /home/admin/.ssh/id_rsa -N ""
-cp /home/admin/.ssh/id_rsa.pub /home/admin/.ssh/authorized_keys
+# Generate SSH key and set up shared authorized_keys
+if [ ! -f /home/admin/.ssh/id_rsa ]; then
+    ssh-keygen -t rsa -f /home/admin/.ssh/id_rsa -N ""
+    cp /home/admin/.ssh/id_rsa.pub /home/admin/.ssh/authorized_keys
+    chown -R admin:admin /home/admin/.ssh
+    chmod 700 /home/admin/.ssh
+    chmod 600 /home/admin/.ssh/authorized_keys
+fi
+
+touch /tmp/ssh-ready
 
 tail -f /dev/null
